use config object and move function to iles
